fix: automatic linting and formatting

+++ src/test/Funding.t.sol
@@ -10,14 +10,14 @@ import {FixedPointMathLib} from "solmate/utils/FixedPointMathLib.sol";
 import "../interfaces/erc20/IERC20.sol";
 
 contract FundingTest is BaseFixture {
-    using FixedPointMathLib for uint;
+    using FixedPointMathLib for uint256;
 
     function setUp() public override {
         BaseFixture.setUp();
     }
 
     function testDiscountRateBasics() public {
-    /*
+        /*
         @fatima: confirm the discount rate is functional
         - access control for setting discount rate (i.e. the proper accounts can call the function and it works. improper accounts revert when attempting to call)
         - access control for setting discount rate limits
@@ -29,9 +29,16 @@ contract FundingTest is BaseFixture {
         fundingCvx.setDiscountLimits(10, 50);
         vm.prank(address(policyOps));
         fundingCvx.setDiscount(20);
-        (uint256 discount,uint256 minDiscount,uint256 maxDiscount,,,) = fundingCvx.funding();
+        (
+            uint256 discount,
+            uint256 minDiscount,
+            uint256 maxDiscount,
+            ,
+            ,
+
+        ) = fundingCvx.funding();
         // check if discount is set
-        assertEq(discount,20);
+        assertEq(discount, 20);
 
         // setting discount above maximum limit
 
@@ -54,7 +61,7 @@ contract FundingTest is BaseFixture {
         // calling with correct role
         vm.prank(address(governance));
         fundingCvx.setDiscountLimits(0, 50);
-        (,minDiscount,maxDiscount,,,) = fundingCvx.funding();
+        (, minDiscount, maxDiscount, , , ) = fundingCvx.funding();
 
         // checking if limits are set
         assertEq(minDiscount, 0);
@@ -71,15 +78,34 @@ contract FundingTest is BaseFixture {
         fundingCvx.setDiscountLimits(0, 20);
     }
 
-    function testDiscountRateBuysCvx(uint256 assetAmountIn, uint32 discount, uint256 citadelPrice) public {
-        _testDiscountRateBuys(fundingCvx, cvx, assetAmountIn, discount, citadelPrice);
-
+    function testDiscountRateBuysCvx(
+        uint256 assetAmountIn,
+        uint32 discount,
+        uint256 citadelPrice
+    ) public {
+        _testDiscountRateBuys(
+            fundingCvx,
+            cvx,
+            assetAmountIn,
+            discount,
+            citadelPrice
+        );
     }
 
-    function testDiscountRateBuysWbtc(uint256 assetAmountIn, uint32 discount, uint256 citadelPrice) public {
+    function testDiscountRateBuysWbtc(
+        uint256 assetAmountIn,
+        uint32 discount,
+        uint256 citadelPrice
+    ) public {
         // wBTC is an 8 decimal example
         // TODO: Fix comparator calls in inner function as per that functions comment
-        _testDiscountRateBuys(fundingWbtc, wbtc, assetAmountIn, discount, citadelPrice);
+        _testDiscountRateBuys(
+            fundingWbtc,
+            wbtc,
+            assetAmountIn,
+            discount,
+            citadelPrice
+        );
     }
 
     function testSetAssetCap() public {
@@ -90,11 +116,11 @@ contract FundingTest is BaseFixture {
         // setting asset cap from correct account
         vm.prank(policyOps);
         fundingCvx.setAssetCap(1000e18);
-        (,,,,, uint256 assetCap) = fundingCvx.funding();
+        (, , , , , uint256 assetCap) = fundingCvx.funding();
         assertEq(assetCap, 1000e18); // check if assetCap is set
 
         // checking assetCap can not be less than accumulated funds.
-         _testDiscountRateBuys(fundingCvx, cvx, 100e18, 3000, 100e18);
+        _testDiscountRateBuys(fundingCvx, cvx, 100e18, 3000, 100e18);
         vm.prank(policyOps);
         vm.expectRevert("cannot decrease cap below global sum of assets in");
         fundingCvx.setAssetCap(10e18);
@@ -108,7 +134,7 @@ contract FundingTest is BaseFixture {
         // Increasing cap allows for deposit
         vm.prank(policyOps);
         fundingCvx.setAssetCap(10000e18);
-        (,,,,, assetCap) = fundingCvx.funding();
+        (, , , , , assetCap) = fundingCvx.funding();
         assertEq(assetCap, 10000e18); // check if assetCap is set
 
         uint256 citadelAmountOutExpected = fundingCvx.getAmountOut(1000e18);
@@ -122,7 +148,6 @@ contract FundingTest is BaseFixture {
     }
 
     function testClaimAssetToTreasury() public {
-
         vm.prank(address(1));
         vm.expectRevert("GAC: invalid-caller-role");
         fundingCvx.claimAssetToTreasury();
@@ -154,8 +179,7 @@ contract FundingTest is BaseFixture {
         balanceAfter = cvx.balanceOf(fundingCvx.saleRecipient());
 
         // difference should be equal to amount transferred
-        assertEq(balanceAfter-balanceBefore, amount); 
-
+        assertEq(balanceAfter - balanceBefore, amount);
     }
 
     function testSweep() public {
@@ -171,7 +195,9 @@ contract FundingTest is BaseFixture {
         // Check that the funding asset can't be sweeped
         erc20utils.forceMintTo(address(fundingCvx), address(cvx), 100e18);
 
-        vm.expectRevert("cannot sweep funding asset, use claimAssetToTreasury()");
+        vm.expectRevert(
+            "cannot sweep funding asset, use claimAssetToTreasury()"
+        );
         fundingCvx.sweep(address(cvx));
 
         vm.expectRevert("nothing to sweep");
@@ -181,17 +207,15 @@ contract FundingTest is BaseFixture {
 
         address saleRecipient = fundingCvx.saleRecipient();
 
-        uint saleRecipientWbtcBefore = wbtc.balanceOf(saleRecipient);
+        uint256 saleRecipientWbtcBefore = wbtc.balanceOf(saleRecipient);
 
         fundingCvx.sweep(address(wbtc));
 
-        uint saleRecipientWbtcAfter = wbtc.balanceOf(saleRecipient);
+        uint256 saleRecipientWbtcAfter = wbtc.balanceOf(saleRecipient);
 
-        assertEq(saleRecipientWbtcAfter-saleRecipientWbtcBefore, 100e8);
+        assertEq(saleRecipientWbtcAfter - saleRecipientWbtcBefore, 100e8);
 
         vm.stopPrank();
-
-
     }
 
     function testAccessControl() public {
@@ -203,7 +227,7 @@ contract FundingTest is BaseFixture {
         // setting discountManager from correct account
         vm.prank(governance);
         fundingCvx.setDiscountManager(address(2));
-        (,,,address discountManager,,) = fundingCvx.funding();
+        (, , , address discountManager, , ) = fundingCvx.funding();
         assertEq(discountManager, address(2)); // check if discountManager is set
 
         vm.prank(address(1));
@@ -233,13 +257,17 @@ contract FundingTest is BaseFixture {
         fundingCvx.setSaleRecipient(address(0));
     }
 
-    function testDepositModifiers() public{
+    function testDepositModifiers() public {
         // flagging citadelPriceFlag should freeze deposit
         vm.prank(governance);
         fundingCvx.setCitadelAssetPriceBounds(0, 5000);
         vm.prank(eoaOracle);
         fundingCvx.updateCitadelPriceInAsset(6000);
-        vm.expectRevert(bytes("Funding: citadel price from oracle flagged and pending review"));
+        vm.expectRevert(
+            bytes(
+                "Funding: citadel price from oracle flagged and pending review"
+            )
+        );
         fundingCvx.deposit(10e18, 0);
     }
 
@@ -250,16 +278,21 @@ contract FundingTest is BaseFixture {
         uint32 _discount,
         uint256 _citadelPrice
     ) public {
-
         emit log_named_uint("Asset Amount in", _assetAmountIn);
         emit log_named_uint("Discount", _discount);
         emit log_named_uint("Citadel Price", _citadelPrice);
 
         // discount < MAX_BPS = 10000
-        vm.assume(_discount<10000 && _assetAmountIn>0 && _citadelPrice>0 && _assetAmountIn<1000000000e18 && _citadelPrice<1000000000e18);
+        vm.assume(
+            _discount < 10000 &&
+                _assetAmountIn > 0 &&
+                _citadelPrice > 0 &&
+                _assetAmountIn < 1000000000e18 &&
+                _citadelPrice < 1000000000e18
+        );
 
         // Adjust funding cap as needed
-        (,,,,, uint256 assetCap) = fundingContract.funding();
+        (, , , , , uint256 assetCap) = fundingContract.funding();
         if (_assetAmountIn > assetCap) {
             vm.prank(policyOps);
             fundingContract.setAssetCap(_assetAmountIn);
@@ -274,7 +307,9 @@ contract FundingTest is BaseFixture {
         vm.prank(eoaOracle);
         fundingContract.updateCitadelPriceInAsset(_citadelPrice); // set citadel price
 
-        uint256 citadelAmountOutExpected = fundingContract.getAmountOut(_assetAmountIn);
+        uint256 citadelAmountOutExpected = fundingContract.getAmountOut(
+            _assetAmountIn
+        );
 
         vm.prank(governance);
         citadel.mint(address(fundingContract), citadelAmountOutExpected); // fundingContract should have citadel to transfer to user
@@ -288,7 +323,7 @@ contract FundingTest is BaseFixture {
         // getAmountsOut returns 0 if (amountsIn * price) < 1x10^(decimals)), hence depositFor reverts.
         // This is acceptable since the price has lower bounds. The transaction will revert if the user
         // attempts to deposit the extremely small amounts that would trigger this behavior.
-        if(citadelAmountOutExpected == 0) {
+        if (citadelAmountOutExpected == 0) {
             vm.expectRevert("Amount 0");
         }
 
@@ -305,43 +340,82 @@ contract FundingTest is BaseFixture {
 
         // Checks (Note: xCTDL 1:1 CTDL at the beginning)
         assertEq(comparator.diff("citadel.balanceOf(shrimp)"), 0);
-        assertEq(comparator.diff("xCitadel.balanceOf(shrimp)"), citadelAmountOutExpected);
+        assertEq(
+            comparator.diff("xCitadel.balanceOf(shrimp)"),
+            citadelAmountOutExpected
+        );
 
-        if (keccak256(abi.encodePacked(token.symbol())) == keccak256(abi.encodePacked(("CVX")))) {
-            assertEq(comparator.negDiff("citadel.balanceOf(fundingCvx)"), citadelAmountOutExpected);
-            assertEq(comparator.diff("cvx.balanceOf(treasuryVault)"), _assetAmountIn);
-            assertEq(comparator.negDiff("cvx.balanceOf(shrimp)"), _assetAmountIn);
+        if (
+            keccak256(abi.encodePacked(token.symbol())) ==
+            keccak256(abi.encodePacked(("CVX")))
+        ) {
+            assertEq(
+                comparator.negDiff("citadel.balanceOf(fundingCvx)"),
+                citadelAmountOutExpected
+            );
+            assertEq(
+                comparator.diff("cvx.balanceOf(treasuryVault)"),
+                _assetAmountIn
+            );
+            assertEq(
+                comparator.negDiff("cvx.balanceOf(shrimp)"),
+                _assetAmountIn
+            );
         } else {
-            assertEq(comparator.negDiff("citadel.balanceOf(fundingWbtc)"), citadelAmountOutExpected);
-            assertEq(comparator.diff("wbtc.balanceOf(treasuryVault)"), _assetAmountIn);
-            assertEq(comparator.negDiff("wbtc.balanceOf(shrimp)"), _assetAmountIn);
+            assertEq(
+                comparator.negDiff("citadel.balanceOf(fundingWbtc)"),
+                citadelAmountOutExpected
+            );
+            assertEq(
+                comparator.diff("wbtc.balanceOf(treasuryVault)"),
+                _assetAmountIn
+            );
+            assertEq(
+                comparator.negDiff("wbtc.balanceOf(shrimp)"),
+                _assetAmountIn
+            );
         }
 
         // check citadelAmoutOut is same as expected
         assertEq(citadelAmountOut, citadelAmountOutExpected);
         assertEq(xCitadel.balanceOf(address(fundingContract)), 0);
-
     }
 
-    function _testBuy(Funding fundingContract, uint assetIn, uint citadelPrice) internal {
+    function _testBuy(
+        Funding fundingContract,
+        uint256 assetIn,
+        uint256 citadelPrice
+    ) internal {
         // just make citadel appear rather than going through minting flow here
-        erc20utils.forceMintTo(address(fundingContract), address(citadel), 100000e18);
+        erc20utils.forceMintTo(
+            address(fundingContract),
+            address(citadel),
+            100000e18
+        );
 
         vm.prank(eoaOracle);
 
         // CVX funding contract gives us a 18 decimal example
         fundingContract.updateCitadelPriceInAsset(citadelPrice);
 
-        uint expectedAssetOut = assetIn.divWadUp(citadelPrice);
+        uint256 expectedAssetOut = assetIn.divWadUp(citadelPrice);
 
-        uint256 citadelAmountOutExpected = fundingContract.getAmountOut(assetIn);
+        uint256 citadelAmountOutExpected = fundingContract.getAmountOut(
+            assetIn
+        );
 
         emit log_named_uint("Citadel Price", citadelPrice);
 
         vm.startPrank(whale);
 
-        require(cvx.balanceOf(whale) >= assetIn, "buyer has insufficent assets for specified buy amount");
-        require(citadel.balanceOf(address(fundingContract)) >= expectedAssetOut, "funding has insufficent citadel for specified buy amount");
+        require(
+            cvx.balanceOf(whale) >= assetIn,
+            "buyer has insufficent assets for specified buy amount"
+        );
+        require(
+            citadel.balanceOf(address(fundingContract)) >= expectedAssetOut,
+            "funding has insufficent citadel for specified buy amount"
+        );
 
         comparator.snapPrev();
         cvx.approve(address(fundingContract), cvx.balanceOf(whale));
@@ -351,13 +425,19 @@ contract FundingTest is BaseFixture {
 
         // user trades in asset for citadel in xCitadel form.
         assertEq(comparator.diff("citadel.balanceOf(whale)"), 0);
-        assertEq(comparator.diff("xCitadel.balanceOf(whale)"), expectedAssetOut);
+        assertEq(
+            comparator.diff("xCitadel.balanceOf(whale)"),
+            expectedAssetOut
+        );
         assertEq(comparator.negDiff("cvx.balanceOf(whale)"), assetIn);
 
         // funding contract loses citadel and sends asset to saleRecipient. should never hold a xCitadel balance (deposited for each user) (gas costs?)
 
         // TODO: Improve comparator to easily add new entity for all balance calls.
-        assertEq(comparator.negDiff("citadel.balanceOf(fundingCvx)"), expectedAssetOut);
+        assertEq(
+            comparator.negDiff("citadel.balanceOf(fundingCvx)"),
+            expectedAssetOut
+        );
         assertEq(comparator.diff("cvx.balanceOf(treasuryVault)"), assetIn);
 
         assertEq(xCitadel.balanceOf(address(fundingContract)), 0);
@@ -365,4 +445,4 @@ contract FundingTest is BaseFixture {
 
         vm.stopPrank();
     }
- }
+}
