+++ src/test/MintAndDistribute.t.sol
@@ -29,19 +29,25 @@ contract MintAndDistributeTest is BaseFixture {
             There unfortunately is the daily manual step of the initial mint destination propotions, we can automate this via contract with some work and oracles.
         */
 
-        assertTrue(address(citadelMinter.supplySchedule()) == address(schedule));
+        assertTrue(
+            address(citadelMinter.supplySchedule()) == address(schedule)
+        );
 
-        uint fundingBps = 4000;
-        uint stakingBps = 3500;
-        uint lockingBps = 2500;
-        uint MAX_BPS = 10000;
+        uint256 fundingBps = 4000;
+        uint256 stakingBps = 3500;
+        uint256 lockingBps = 2500;
+        uint256 MAX_BPS = 10000;
 
-        uint wbtcFundingPoolWeight = 8000;
-        uint cvxFundingPoolWeight = 2000;
-        uint expectedTotalPoolWeight = 10000;
+        uint256 wbtcFundingPoolWeight = 8000;
+        uint256 cvxFundingPoolWeight = 2000;
+        uint256 expectedTotalPoolWeight = 10000;
 
         vm.startPrank(policyOps);
-        citadelMinter.setCitadelDistributionSplit(fundingBps,stakingBps,lockingBps);
+        citadelMinter.setCitadelDistributionSplit(
+            fundingBps,
+            stakingBps,
+            lockingBps
+        );
         // confirm only policy ops can call
         // bps between three positions must add up to 10000 (100%)
 
@@ -49,7 +55,7 @@ contract MintAndDistributeTest is BaseFixture {
         assertTrue(schedule.globalStartTimestamp() == 0);
         vm.expectRevert("SupplySchedule: minting not started");
         citadelMinter.mintAndDistribute();
-        
+
         // policy ops should not be able to start minting schedule
         vm.expectRevert("GAC: invalid-caller-role");
         schedule.setMintingStart(block.timestamp);
@@ -68,7 +74,9 @@ contract MintAndDistributeTest is BaseFixture {
         assertTrue(schedule.globalStartTimestamp() == block.timestamp);
 
         // Attempt to initializeLastMintTimestamp with after already initializing lastMintTimestamp
-        vm.expectRevert("CitadelMinter: last mint timestamp already initialized");
+        vm.expectRevert(
+            "CitadelMinter: last mint timestamp already initialized"
+        );
         citadelMinter.initializeLastMintTimestamp();
 
         vm.stopPrank();
@@ -83,43 +91,77 @@ contract MintAndDistributeTest is BaseFixture {
         vm.expectRevert("CitadelMinter: no funding pools");
         citadelMinter.mintAndDistribute();
 
-        citadelMinter.setFundingPoolWeight(address(fundingWbtc), wbtcFundingPoolWeight);
-        citadelMinter.setFundingPoolWeight(address(fundingCvx), cvxFundingPoolWeight);
+        citadelMinter.setFundingPoolWeight(
+            address(fundingWbtc),
+            wbtcFundingPoolWeight
+        );
+        citadelMinter.setFundingPoolWeight(
+            address(fundingCvx),
+            cvxFundingPoolWeight
+        );
 
-        uint xCitadelBalanceBefore = xCitadel.balance();
+        uint256 xCitadelBalanceBefore = xCitadel.balance();
         comparator.snapPrev();
-        uint expectedMint = schedule.getMintable(citadelMinter.lastMintTimestamp());
+        uint256 expectedMint = schedule.getMintable(
+            citadelMinter.lastMintTimestamp()
+        );
 
         citadelMinter.mintAndDistribute();
 
         comparator.snapCurr();
 
         // funding pools should recieve based on funding bps and pool weights
-        uint expectedToFunding = expectedMint * fundingBps / MAX_BPS;
+        uint256 expectedToFunding = (expectedMint * fundingBps) / MAX_BPS;
 
-        uint totalPoolWeight = citadelMinter.totalFundingPoolWeight();
+        uint256 totalPoolWeight = citadelMinter.totalFundingPoolWeight();
         assertTrue(totalPoolWeight == expectedTotalPoolWeight);
 
-        uint expectedToWbtcFunding = expectedToFunding * wbtcFundingPoolWeight / totalPoolWeight;
-        assertEq(comparator.diff("citadel.balanceOf(fundingWbtc)"), expectedToWbtcFunding);
+        uint256 expectedToWbtcFunding = (expectedToFunding *
+            wbtcFundingPoolWeight) / totalPoolWeight;
+        assertEq(
+            comparator.diff("citadel.balanceOf(fundingWbtc)"),
+            expectedToWbtcFunding
+        );
 
-        uint expectedToCvxFunding = expectedToFunding * cvxFundingPoolWeight / totalPoolWeight;
-        assertEq(comparator.diff("citadel.balanceOf(fundingCvx)"), expectedToCvxFunding);
+        uint256 expectedToCvxFunding = (expectedToFunding *
+            cvxFundingPoolWeight) / totalPoolWeight;
+        assertEq(
+            comparator.diff("citadel.balanceOf(fundingCvx)"),
+            expectedToCvxFunding
+        );
 
         // staking ppfs should increase based on staking bps
-        uint expectedToStakers = expectedMint * stakingBps / MAX_BPS;
+        uint256 expectedToStakers = (expectedMint * stakingBps) / MAX_BPS;
 
-        emit log_named_uint("xCitadel ppfs before", comparator.prev("xCitadel.getPricePerFullShare()"));
-        emit log_named_uint("xCitadel ppfs after", comparator.curr("xCitadel.getPricePerFullShare()"));
-        emit log_named_uint("change in xCitadel ppfs", comparator.diff("xCitadel.getPricePerFullShare()"));
+        emit log_named_uint(
+            "xCitadel ppfs before",
+            comparator.prev("xCitadel.getPricePerFullShare()")
+        );
+        emit log_named_uint(
+            "xCitadel ppfs after",
+            comparator.curr("xCitadel.getPricePerFullShare()")
+        );
+        emit log_named_uint(
+            "change in xCitadel ppfs",
+            comparator.diff("xCitadel.getPricePerFullShare()")
+        );
 
-        emit log_named_uint("xCitadel total supply before", comparator.prev("xCitadel.totalSupply()"));
-        emit log_named_uint("xCitadel total supply after", comparator.curr("xCitadel.totalSupply()"));
-        emit log_named_uint("xCitadel change in supply", comparator.diff("xCitadel.totalSupply()"));
+        emit log_named_uint(
+            "xCitadel total supply before",
+            comparator.prev("xCitadel.totalSupply()")
+        );
+        emit log_named_uint(
+            "xCitadel total supply after",
+            comparator.curr("xCitadel.totalSupply()")
+        );
+        emit log_named_uint(
+            "xCitadel change in supply",
+            comparator.diff("xCitadel.totalSupply()")
+        );
 
         // locking reward schedule should modulate based on locking bps
-        uint expectedToLockers = expectedMint * lockingBps / MAX_BPS;
-        uint xCitadelBalanceAfter = xCitadel.balance();
+        uint256 expectedToLockers = (expectedMint * lockingBps) / MAX_BPS;
+        uint256 xCitadelBalanceAfter = xCitadel.balance();
 
         // total supply should increase as the amount is deposited to locker
         assertEq(comparator.diff("xCitadel.totalSupply()"), expectedToLockers);
@@ -127,12 +169,22 @@ contract MintAndDistributeTest is BaseFixture {
         // the difference of total supply and balance should be expectedToStakers
         // expectedToStakers is directly transferred not deposited, which increases balance of citadel but does not affect totalSupply
         // which causes ppfs increase
-        assertEq(xCitadelBalanceAfter - comparator.curr("xCitadel.totalSupply()"), expectedToStakers);
-        assertEq(comparator.diff("xCitadel.getPricePerFullShare()"), (expectedToStakers * 1e18)/comparator.curr("xCitadel.totalSupply()"));
-        
-        // expectedToStakers and expectedToLockers both go to xCitadel so 
+        assertEq(
+            xCitadelBalanceAfter - comparator.curr("xCitadel.totalSupply()"),
+            expectedToStakers
+        );
+        assertEq(
+            comparator.diff("xCitadel.getPricePerFullShare()"),
+            (expectedToStakers * 1e18) /
+                comparator.curr("xCitadel.totalSupply()")
+        );
+
+        // expectedToStakers and expectedToLockers both go to xCitadel so
         // the difference in balance should be equal to sum of both amounts
-        assertEq(xCitadelBalanceAfter - xCitadelBalanceBefore,expectedToStakers + expectedToLockers);
+        assertEq(
+            xCitadelBalanceAfter - xCitadelBalanceBefore,
+            expectedToStakers + expectedToLockers
+        );
         vm.stopPrank();
     }
 }
